per process net perf
The psutil process io stats comes from /proc/PID/net/netstat which is
global byte counts for all the interfaces, not per process stats.  The
way other utilies get per-process stats is to map sockets to PIDs with
the process file descriptors and then listen on the network for packets
matching those sockets.  This implementation uses gopacket to help with
the packet collecting and filtering.  For now this is only TCP
performance.  Will add UDP and other in the future maybe.